Refactored codebase to split UI layout code
